Move all RMC-specific code to RMC folder
