--- Content.Shared/Actions/Events/SoundActionEvent.cs
@@ -1,3 +1,0 @@
-namespace Content.Shared.Actions.Events;
-
-public sealed partial class SoundActionEvent : InstantActionEvent;
