+++ Content.Shared/_RMC14/Sound/CMSoundSystem.cs
@@ -1,5 +1,6 @@
 ﻿using Content.Shared.Mobs;
 using Content.Shared.Mobs.Systems;
+using Content.Shared.Sound;
 using Robust.Shared.Audio.Systems;
 using Robust.Shared.Network;
 using Robust.Shared.Random;
@@ -10,6 +11,7 @@ namespace Content.Shared._RMC14.Sound;
 public sealed class CMSoundSystem : EntitySystem
 {
     [Dependency] private readonly SharedAudioSystem _audio = default!;
+    [Dependency] private readonly SharedEmitSoundSystem _emitSound = default!;
     [Dependency] private readonly MobStateSystem _mobState = default!;
     [Dependency] private readonly INetManager _net = default!;
     [Dependency] private readonly IRobustRandom _random = default!;
@@ -20,6 +22,8 @@ public sealed class CMSoundSystem : EntitySystem
         SubscribeLocalEvent<RandomSoundComponent, MapInitEvent>(OnRandomMapInit);
 
         SubscribeLocalEvent<SoundOnDeathComponent, MobStateChangedEvent>(OnDeathMobStateChanged);
+
+        SubscribeLocalEvent<EmitSoundOnActionComponent, SoundActionEvent>(OnEmitSoundOnAction);
     }
 
     private void OnRandomMapInit(Entity<RandomSoundComponent> ent, ref MapInitEvent args)
@@ -40,6 +44,14 @@ public sealed class CMSoundSystem : EntitySystem
 
         if (_net.IsServer)
             _audio.PlayPvs(ent.Comp.Sound, ent);
+    }
+
+    private void OnEmitSoundOnAction(Entity<EmitSoundOnActionComponent> ent, ref SoundActionEvent args)
+    {
+        _emitSound.TryEmitSound(ent, ent, args.Performer);
+
+        if (ent.Comp.Handle)
+            args.Handled = true;
     }
 
     public override void Update(float frameTime)
