+++ Content.Shared/_RMC14/Sound/SoundActionEvent.cs
@@ -1,0 +1,5 @@
+using Content.Shared.Actions;
+
+namespace Content.Shared._RMC14.Sound;
+
+public sealed partial class SoundActionEvent : InstantActionEvent;
