+++ Content.Shared/_RMC14/Whistle/RMCWhistleSystem.cs
@@ -1,3 +1,4 @@
+using Content.Shared._RMC14.Sound;
 using Content.Shared.Actions;
 using Content.Shared.Actions.Events;
 using Content.Shared.Interaction.Events;
